house-hunting-app/requirements.txt
# ── UI ────────────────────────────────────────────────────────────────────────
streamlit
streamlit-sortables

# ── LLM / LangChain stack ─────────────────────────────────────────────────────
langchain
langchain-core
langchain-google-genai
langgraph

# ── Backend API ───────────────────────────────────────────────────────────────
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests

# ── Data / scoring ────────────────────────────────────────────────────────────
numpy

# ── Existing (keep) ───────────────────────────────────────────────────────────
google-genai
